+++ src/config.ts
@@ -1,4 +1,20 @@
 import { PlatformConfig } from 'homebridge';
+import { TuyaDeviceSchemaProperty, TuyaDeviceSchemaType } from './device/TuyaDevice';
+
+export interface TuyaPlatformDeviceSchemaConfig {
+  oldCode: string;
+  code: string;
+  type: TuyaDeviceSchemaType;
+  property: TuyaDeviceSchemaProperty;
+  onGet: string;
+  onSet: string;
+}
+
+export interface TuyaPlatformDeviceConfig {
+  id: string;
+  category: string;
+  schema: Array<TuyaPlatformDeviceSchemaConfig>;
+}
 
 export interface TuyaPlatformCustomConfigOptions {
   projectType: '1';
@@ -7,6 +23,7 @@ export interface TuyaPlatformCustomConfigOptions {
   accessKey: string;
   username: string;
   password: string;
+  deviceOverrides: Array<TuyaPlatformDeviceConfig>;
 }
 
 export interface TuyaPlatformHomeConfigOptions {
@@ -18,7 +35,7 @@ export interface TuyaPlatformHomeConfigOptions {
   password: string;
   appSchema: string;
   homeWhitelist: Array<number>;
-  sceneWhitelist: Array<string>;
+  deviceOverrides: Array<TuyaPlatformDeviceConfig>;
 }
 
 export type TuyaPlatformConfigOptions = TuyaPlatformCustomConfigOptions | TuyaPlatformHomeConfigOptions;
@@ -32,6 +49,7 @@ export const customOptionsSchema = {
     endpoint: { type: 'string', format: 'url', required: true },
     accessId: { type: 'string', required: true },
     accessKey: { type: 'string', required: true },
+    deviceOverrides: { 'type': 'array' },
   },
 };
 
@@ -44,6 +62,6 @@ export const homeOptionsSchema = {
     password: { type: 'string', required: true },
     appSchema: { 'type': 'string', required: true },
     homeWhitelist: { 'type': 'array' },
-    sceneWhitelist: { 'type': 'array' },
+    deviceOverrides: { 'type': 'array' },
   },
 };