+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### Added
 - Add scene support. (#118)
-- Add `sceneWhitelist` option for whitelisting scenes. (#118)
 - Add Wireless Switch support (`wxkg`).
 - Add Solar Light support (`tyndj`).
 - Add Dehumidifier support (`cs`).
 - Add Scene Switch support (`wxkg`).
+- Add device overriding config support. "non-standard" devices have possibility to be supported now.
+
 
 ### Changed
 - Support Ceiling Fan icon customize and Floor Fan `lock`, `swing` feature. (#131)